+++ docs/src/pages/company/about/about.md
@@ -1,21 +1,23 @@
 # About Us
 
-<p class="description">We’re on a mission to make React application development simple.</p>
+<p class="description">We’re on a mission to make building UIs with React fun.</p>
 
-## Our mission
+## Our Mission
 
-Our company is focused on making React application development easier, better, and accessible to more people.
+Our company is focused on making React UIs development easier, better, and accessible to more people.
 We build open source and commercial tools used by many hundreds of thousands of developers in production.
 We're proud not only of the products we make, but also the community and partnerships we've cultivated with other developers and companies.
 
-## Our vision
+## Our Vision
 
-Our vision is to provide React components for faster and easier web development, so you can build your own design system, or start with Material Design. [Read more](/discover-more/vision/).
+Our vision is to provide React components for faster and easier web development. [Read more](/discover-more/vision/).
 
-## Values
+## Our Values
 
 Our core values include transparency (our work is public most of the time); creating a safe, high-trust team; building incredible developer experiences; maintaining a healthy working environment; and helping to deliver web experiences that feel amazing to use on every device and connection type.
 
-## History
+## Our Story
 
-Material-UI started back in 2014 to unify React and Material Design. Today, Material-UI has grown to become one of the world's most popular React UI libraries – backed by a vibrant community of more than 1M developers in over 180 countries.
+Material-UI started back in 2014 to unify React and Material Design.
+
+Today, Material-UI has grown to become one of the world's most popular React UI libraries – backed by a vibrant community of more than 1M developers in over 180 countries.
